+++ pymilo/chains/ensemble_chain.py
@@ -14,8 +14,8 @@ from ..pymilo_param import SKLEARN_ENSEMBLE_TABLE
 
 from .util import get_concrete_transporter
 
-from ..exceptions.serialize_exception import PymiloSerializationException, SerilaizatoinErrorTypes
-from ..exceptions.deserialize_exception import PymiloDeserializationException, DeSerilaizatoinErrorTypes
+from ..exceptions.serialize_exception import PymiloSerializationException, SerializationErrorTypes
+from ..exceptions.deserialize_exception import PymiloDeserializationException, DeserializationErrorTypes
 
 from ..utils.util import get_sklearn_type, check_str_in_iterable
 
@@ -89,7 +89,7 @@ def transport_ensemble(request, command, is_inner_model=False):
         except Exception as e:
             raise PymiloSerializationException(
                 {
-                    'error_type': SerilaizatoinErrorTypes.VALID_MODEL_INVALID_INTERNAL_STRUCTURE,
+                    'error_type': SerializationErrorTypes.VALID_MODEL_INVALID_INTERNAL_STRUCTURE,
                     'error': {
                         'Exception': repr(e),
                         'Traceback': format_exc(),
@@ -97,13 +97,13 @@ def transport_ensemble(request, command, is_inner_model=False):
                     'object': request,
                 })
 
-    elif command == Command.DESERIALZIE:
+    elif command == Command.DESERIALIZE:
         try:
             return deserialize_ensemble(request, is_inner_model)
         except Exception as e:
             raise PymiloDeserializationException(
                 {
-                    'error_type': SerilaizatoinErrorTypes.VALID_MODEL_INVALID_INTERNAL_STRUCTURE,
+                    'error_type': SerializationErrorTypes.VALID_MODEL_INVALID_INTERNAL_STRUCTURE,
                     'error': {
                         'Exception': repr(e),
                         'Traceback': format_exc()},
@@ -145,7 +145,7 @@ def deserialize_possible_ml_model(possible_serialized_ml_model):
         return True, transporter({
             "data": possible_serialized_ml_model["pymilo-inner-model-data"],
             "type": possible_serialized_ml_model["pymilo-inner-model-type"]
-        }, Command.DESERIALZIE, is_inner_model=True)
+        }, Command.DESERIALIZE, is_inner_model=True)
     else:
         return False, possible_serialized_ml_model
 
@@ -321,7 +321,7 @@ def deserialize_ensemble(ensemble, is_inner_model=False):
     for transporter in ENSEMBLE_CHAIN:
         if transporter != "GeneralDataStructureTransporter":
             ENSEMBLE_CHAIN[transporter].transport(
-                ensemble, Command.DESERIALZIE, is_inner_model)
+                ensemble, Command.DESERIALIZE, is_inner_model)
 
     for key, value in data.items():
         if isinstance(value, dict):
@@ -354,7 +354,7 @@ def deserialize_ensemble(ensemble, is_inner_model=False):
         if has_ml_model:
             data[key] = result
 
-    ENSEMBLE_CHAIN["GeneralDataStructureTransporter"].transport(ensemble, Command.DESERIALZIE, is_inner_model)
+    ENSEMBLE_CHAIN["GeneralDataStructureTransporter"].transport(ensemble, Command.DESERIALIZE, is_inner_model)
 
     _type = None
     raw_model = None
@@ -393,17 +393,17 @@ def _validate_input(model, command):
         else:
             raise PymiloSerializationException(
                 {
-                    'error_type': SerilaizatoinErrorTypes.INVALID_MODEL,
+                    'error_type': SerializationErrorTypes.INVALID_MODEL,
                     'object': model
                 }
             )
-    elif command == Command.DESERIALZIE:
+    elif command == Command.DESERIALIZE:
         if is_ensemble(model.type):
             return
         else:
             raise PymiloDeserializationException(
                 {
-                    'error_type': DeSerilaizatoinErrorTypes.INVALID_MODEL,
+                    'error_type': DeserializationErrorTypes.INVALID_MODEL,
                     'object': model
                 }
             )