+++ pymilo/chains/svm_chain.py
@@ -7,8 +7,8 @@ from ..transporters.randomstate_transporter import RandomStateTransporter
 from ..transporters.preprocessing_transporter import PreprocessingTransporter
 
 from ..pymilo_param import SKLEARN_SVM_TABLE
-from ..exceptions.serialize_exception import PymiloSerializationException, SerilaizatoinErrorTypes
-from ..exceptions.deserialize_exception import PymiloDeserializationException, DeSerilaizatoinErrorTypes
+from ..exceptions.serialize_exception import PymiloSerializationException, SerializationErrorTypes
+from ..exceptions.deserialize_exception import PymiloDeserializationException, DeserializationErrorTypes
 
 from ..utils.util import get_sklearn_type
 
@@ -56,7 +56,7 @@ def transport_svm(request, command, is_inner_model=False):
         except Exception as e:
             raise PymiloSerializationException(
                 {
-                    'error_type': SerilaizatoinErrorTypes.VALID_MODEL_INVALID_INTERNAL_STRUCTURE,
+                    'error_type': SerializationErrorTypes.VALID_MODEL_INVALID_INTERNAL_STRUCTURE,
                     'error': {
                         'Exception': repr(e),
                         'Traceback': format_exc(),
@@ -64,13 +64,13 @@ def transport_svm(request, command, is_inner_model=False):
                     'object': request,
                 })
 
-    elif command == Command.DESERIALZIE:
+    elif command == Command.DESERIALIZE:
         try:
             return deserialize_svm(request, is_inner_model)
         except Exception as e:
             raise PymiloDeserializationException(
                 {
-                    'error_type': SerilaizatoinErrorTypes.VALID_MODEL_INVALID_INTERNAL_STRUCTURE,
+                    'error_type': SerializationErrorTypes.VALID_MODEL_INVALID_INTERNAL_STRUCTURE,
                     'error': {
                         'Exception': repr(e),
                         'Traceback': format_exc()},
@@ -112,7 +112,7 @@ def deserialize_svm(svm, is_inner_model=False):
 
     for transporter in SVM_CHAIN:
         SVM_CHAIN[transporter].transport(
-            svm, Command.DESERIALZIE, is_inner_model)
+            svm, Command.DESERIALIZE, is_inner_model)
     for item in data:
         setattr(raw_model, item, data[item])
     return raw_model
@@ -134,17 +134,17 @@ def _validate_input(model, command):
         else:
             raise PymiloSerializationException(
                 {
-                    'error_type': SerilaizatoinErrorTypes.INVALID_MODEL,
+                    'error_type': SerializationErrorTypes.INVALID_MODEL,
                     'object': model
                 }
             )
-    elif command == Command.DESERIALZIE:
+    elif command == Command.DESERIALIZE:
         if is_svm(model.type):
             return
         else:
             raise PymiloDeserializationException(
                 {
-                    'error_type': DeSerilaizatoinErrorTypes.INVALID_MODEL,
+                    'error_type': DeserializationErrorTypes.INVALID_MODEL,
                     'object': model
                 }
             )