+++ pymilo/transporters/treepredictor_transporter.py
@@ -11,7 +11,7 @@ class TreePredictorTransporter(AbstractTransporter):
 
     def serialize(self, data, key, model_type):
         """
-        Serialize TreePredictor object[usefull in HistGradientBoosting(Regressor,Classifier)].
+        Serialize TreePredictor object[useful in HistGradientBoosting(Regressor,Classifier)].
 
         serialize the data[key] of the given model which type is model_type.
         basically in order to fully serialize a model, we should traverse over all the keys of its data dictionary and
@@ -33,7 +33,7 @@ class TreePredictorTransporter(AbstractTransporter):
 
     def deserialize(self, data, key, model_type):
         """
-        Deserialize previously pymilo serialized TreePredictor object[usefull in HistGradientBoosting(Regressor,Classifier)].
+        Deserialize previously pymilo serialized TreePredictor object[useful in HistGradientBoosting(Regressor,Classifier)].
 
         deserialize the data[key] of the given model which type is model_type.
         basically in order to fully deserialize a model, we should traverse over all the keys of its serialized data dictionary and